requirements.txt
# Core HTTP / retry
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1

# WebSocket (CLOB WSS market channel)
websockets==14.1

# Data validation / config
pydantic==2.9.2

# Postgres
psycopg[binary]==3.2.3

# Polymarket CLOB (paper trading will primarily use data endpoints; trading endpoints not required)
py-clob-client==0.19.0

# Optional: structured logging
structlog==24.4.0


